Add condition before message delete. Handle the bgsbot non tick message. change print to log in a few places until I can figure out the docker flush issue

+++ spyplane/listeners.py
@@ -21,25 +21,25 @@ class Listeners:
 
 @bot.event
 async def on_ready():
-    print(f'{bot.user.name} has connected to Discord server. Version: {__version__}')
+    log(f'{bot.user.name} has connected to Discord server. Version: {__version__}')
     bot.channel = bot.get_channel(CONTROL_CHANNEL)
     bot.lock = asyncio.Lock()
     emoji = bot.get_emoji(EMOJI_BULLSEYE)
     bot.emoji_bullseye = emoji or '✅'
-    await bot.channel.send(f'{bot.user.name} has connected to Discord server. Version: {__version__}')
+    # await bot.channel.send(f'{bot.user.name} has connected to Discord server. Version: {__version__}')
 
 
 @bot.event
 async def on_disconnect():
-    print(f'Spy Plane has disconnected from discord server. Version: {__version__}.')
+    log(f'Spy Plane has disconnected from discord server. Version: {__version__}.')
 
 
 @bot.event
 async def on_error(event, *args, **kwargs):
-    print("ERROR")
-    print(event)
-    print(args)
-    print(kwargs)
+    log("ERROR")
+    log(event)
+    log(args)
+    log(kwargs)
 
 
 record = ScoutRecordingService()
@@ -47,25 +47,30 @@ record = ScoutRecordingService()
 
 @bot.event
 async def on_raw_reaction_add(payload: RawReactionActionEvent):
-    if payload.channel_id!=CONTROL_CHANNEL:
-        # print(f"Not the right channel {payload.channel_id}")
-        return
-    if payload.user_id==bot.user.id:
-        # print(f"Not the right user {payload.user_id}")
-        return
-    if str(payload.emoji)!=str(bot.emoji_bullseye):
-        print(f"Not the target emoji {payload.emoji} {bot.emoji_bullseye} {payload.emoji.name} {bot.emoji_bullseye.name}")
-        return
-    message: Message = await bot.channel.fetch_message(payload.message_id)
-    asyncio.create_task(record.record_reaction(message.content, payload.member.name, payload.member.id))  # Another option is to try a Queue
-    await message.delete()
+    try:
+        if payload.channel_id!=CONTROL_CHANNEL:
+            # log(f"Not the right channel {payload.channel_id}")
+            return
+        if payload.user_id==bot.user.id:
+            # log(f"Not the right user {payload.user_id}")
+            return
+        if str(payload.emoji)!=str(bot.emoji_bullseye):
+            log(f"Not the target emoji {payload.emoji} {bot.emoji_bullseye} {payload.emoji.name} {bot.emoji_bullseye.name}")
+            return
+        message: Message = await bot.channel.fetch_message(payload.message_id)
+        asyncio.create_task(record.record_reaction(message.content, payload.member.name, payload.member.id))  # Another option is to try a Queue
+        if not message.pinned: # prevent deleting pinned messages with reactions in the channel
+            await message.delete()
+    except Exception as e:
+        log("Exception on reaction add")
+        log(str(e))
 
 
 @bot.event
 async def on_message(msg):
     try:
         if msg.author.id==bot.user.id:
-            # print("Message from bot to itself")
+            # log("Message from bot to itself")
             pass
         elif msg.author.id==BGS_BOT_USER_ID and msg.channel.id==TICK_CHANNEL:
             log("Message from BGS Bot in tick channel!")
